Refactoring

+++ src/main/java/com/example/petproject/model/User.java
@@ -6,10 +6,9 @@ import java.util.Set;
 @Entity
 @Table(name = "usr")
 public class User {
-
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
+    private Long id;
     private String username;
     private String parole;
     private String active;
@@ -19,11 +18,11 @@ public class User {
     @Enumerated(EnumType.STRING)
     private Set<Role> roles;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
